--- src/main/java/com/kr/lg/model/common/sns/kakao/dto/KakaoProfile.java
@@ -1,15 +1,0 @@
-package com.kr.lg.model.common.sns.kakao.dto;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class KakaoProfile {
-
-    private String nickname;
-    private String thumbnailImageUrl;
-    private String profileImageUrl;
-    private String isDefaultImage;
-    private String isDefaultNickname;
-}
