--- src/main/java/com/kr/lg/model/common/sns/kakao/dto/KakaoProperties.java
@@ -1,14 +1,0 @@
-package com.kr.lg.model.common.sns.kakao.dto;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class KakaoProperties {
-
-    private String nickname;
-    private String profileImage;
-    private String thumbnailImage;
-
-}
