+++ src/main/java/com/kr/lg/module/login/model/google/GoogleLoginDto.java
@@ -1,0 +1,26 @@
+package com.kr.lg.module.login.model.google;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class GoogleLoginDto {
+    private String iss;
+    private String azp;
+    private String aud;
+    private String sub;
+    private String email;
+    private String emailVerified;
+    private String atHash;
+    private String name;
+    private String picture;
+    private String givenName;
+    private String familyName;
+    private String locale;
+    private String iat;
+    private String exp;
+    private String alg;
+    private String kid;
+    private String typ;
+}
