+++ src/main/java/com/kr/lg/module/login/model/kakao/dto/KakaoAcount.java
@@ -1,0 +1,18 @@
+package com.kr.lg.module.login.model.kakao.dto;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class KakaoAcount {
+
+    private String profileNicknameNeedsAgreement;
+    private String profileImageNeedsAgreement;
+    private KakaoProfile profile;
+    private String hasEmail;
+    private String emailNeedsAgreement;
+    private String isEmailValid;
+    private String isEmailVerified;
+    private String email;
+}
