+++ src/main/java/com/kr/lg/module/login/model/kakao/dto/KakaoProfile.java
@@ -1,0 +1,15 @@
+package com.kr.lg.module.login.model.kakao.dto;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class KakaoProfile {
+
+    private String nickname;
+    private String thumbnailImageUrl;
+    private String profileImageUrl;
+    private String isDefaultImage;
+    private String isDefaultNickname;
+}
