+++ src/main/java/com/kr/lg/module/login/model/kakao/dto/KakaoProperties.java
@@ -1,0 +1,14 @@
+package com.kr.lg.module.login.model.kakao.dto;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class KakaoProperties {
+
+    private String nickname;
+    private String profileImage;
+    private String thumbnailImage;
+
+}
