--- src/main/java/com/kr/lg/service/sns/KakaoService.java
@@ -1,9 +1,0 @@
-package com.kr.lg.service.sns;
-
-import com.kr.lg.model.common.sns.kakao.KakaoLoginDto;
-import com.kr.lg.model.common.sns.kakao.KakaoLoginRequestDto;
-
-public interface KakaoService {
-
-    KakaoLoginDto kakaoOAuth(KakaoLoginRequestDto requestDto) throws Exception;
-}
