--- src/main/java/com/kr/lg/service/sns/LgService.java
@@ -1,11 +1,0 @@
-package com.kr.lg.service.sns;
-
-import com.kr.lg.model.common.sns.lg.LgLoginDto;
-import org.springframework.transaction.annotation.Transactional;
-
-@Transactional(readOnly = true)
-public interface LgService {
-
-    @Transactional
-    String LgLogin(LgLoginDto lgLoginDto) throws Exception;
-}
