--- src/main/java/com/kr/lg/service/sns/NaverService.java
@@ -1,9 +1,0 @@
-package com.kr.lg.service.sns;
-
-import com.kr.lg.model.common.sns.naver.NaverLoginDto;
-import com.kr.lg.model.common.sns.naver.NaverLoginRequestDto;
-
-public interface NaverService {
-
-    NaverLoginDto naverOAuth(NaverLoginRequestDto requestDto) throws Exception;
-}
