--- src/main/java/com/kr/lg/service/sns/impl/GoogleServiceImpl.java
@@ -1,63 +1,0 @@
-package com.kr.lg.service.sns.impl;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.PropertyNamingStrategy;
-import com.kr.lg.model.common.sns.google.GoogleLoginDto;
-import com.kr.lg.model.common.sns.google.GoogleLoginRequestDto;
-import com.kr.lg.model.common.sns.google.GoogleLoginResponseDto;
-import com.kr.lg.model.common.sns.google.GoogleProp;
-import com.kr.lg.service.sns.GoogleService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class GoogleServiceImpl implements GoogleService {
-
-    private final GoogleProp googleProp;
-
-
-    @Override
-    public GoogleLoginDto googleOAuth(GoogleLoginRequestDto requestDto) throws Exception {
-        RestTemplate restTemplate = new RestTemplate();
-
-        // Http Header 설정
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        HttpEntity<GoogleLoginRequestDto> httpRequestEntity = new HttpEntity<>(requestDto, headers);
-        ResponseEntity<String> apiResponseJson = restTemplate.postForEntity(googleProp.getGoogleAuthUrl() + "/token", httpRequestEntity, String.class);
-
-        // ObjectMapper를 통해 String to Object로 변환
-        ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.setPropertyNamingStrategy(PropertyNamingStrategy.SNAKE_CASE);
-        objectMapper.setSerializationInclusion(JsonInclude.Include.NON_NULL); // NULL이 아닌 값만 응답받기(NULL인 경우는 생략)
-        GoogleLoginResponseDto googleLoginResponse = objectMapper.readValue(apiResponseJson.getBody(), new TypeReference<GoogleLoginResponseDto>() {});
-
-        // 사용자의 정보는 JWT Token으로 저장되어 있고, Id_Token에 값을 저장한다.
-        String jwtToken = googleLoginResponse.getIdToken();
-
-        // JWT Token을 전달해 JWT 저장된 사용자 정보 확인
-        String requestUrl = UriComponentsBuilder.fromHttpUrl(googleProp.getGoogleAuthUrl() + "/tokeninfo").queryParam("id_token", jwtToken).toUriString();
-
-        String resultJson = restTemplate.getForObject(requestUrl, String.class);
-
-        if(resultJson != null) {
-            GoogleLoginDto userInfoDto = objectMapper.readValue(resultJson, new TypeReference<GoogleLoginDto>() {});
-            return userInfoDto;
-        }
-        else {
-            throw new Exception("Google OAuth failed!");
-        }
-
-
-    }
-}
